Restructure types
