+++ src/types/Errors.ts
@@ -1,0 +1,7 @@
+export type ErrorsState = {
+    errors?: {
+      email?: string[];
+      password?: string[];
+    };
+    message?: null | string;
+  };
